Speed up unwinding header dependencies, parse for include file only once
We do unwind on demand, when we already parsed all header files (and
included .c files). We save all intermediate results, and we do not
need to unwind non used header files

We call unwind also just after reading the c file
(not at every scanner)

Additionally we moved in linux browser the linux kernel exception
file (and before scanning the files, we set the exceptions).