Fix two bugs on kbuild (sort of reverting old changes)
On is an improvement of: 4808367f9e35927656cd7450ebcad6750fec2bde

That commit removed infinite recursion, but it caused missing
many configurations. So now we allow some recursions

The second fix: last year we raised exceptions instead of just
logging. We do it again in case of 'depends on' missing the 'on'