read (and re.subs) only when needed + remove defines on block
re.subs is much more slow then expected, so we read file in parse_header
just when needed, and so also the removing of comments, and mergine line
ending with \). BTW an improvement of API.

On the other hand, we add a new re.subs when parsing blocks, in order to
remove #define. This will removes few cases where the defines are put
in a block, and so parsing errors, and few generic items less on results