Include arch headers only once
This will add two items on results (and remove one generic).

I was expecting more (on the other hand, it was unnoticed until now),
but this will reduce few errors in logs, and possibly cause less trouble
in future

BTW in parser we remove the "loop" on a single value. After tests,
there is no need to increase range of ../ in case of <> headers.
For "" we could reduce range from 3 to 2, but test is cheap, so let's
keep it.