Rename license file, do not upload sources, counts lines
Now that we are in GitLab and GitHub, we put license files
(GPL-2 and GPL-3) in a more standard name: LICENSE and
LICENSE.GPL-3. Because we insert (and later), we put
GPL-2 as "default" license.

We do no build sources tar anymore. The official sources
are now in git, and we distribute on GitLab and GitHub.
[I have some clocal copies, but for now I do no provide
directly the sources. If you think it is a problem,
just email me]

We add also a small function in utils, to get line counts
(copied from utils/rebuild-web.sh, but now we can call
it independently, and we can create a file for every
version automatically).