Convert from optparse to argparse, remove other sql references, logging
A sort of "modernization" of code.

argparse is the "next" optparse. Very similar, but now it handle
"positional arguments" (previous "arguments").

SQL: we found 10 years ago that sqlite was too slow to be useful.
It is a lot better to generate the files, and then to import in a batch.

logging: now we use consistently logging module